Changed RifleItem to WeaponItem.
Made it that a player can hold only 1 WeaponItem in the hotbar. this is expected to be very buggy.

Added a helper method to determine if a player has a WeaponItem in the hotbar.

(unrelated to the mod) added a helpful swear counter written in Gem, i stole it from someone but it had an MIT license so don't worry